scale property of sprites can now be float

+++ Spaxel/src/main/java/code/components/particle/ParticleComponent.java
@@ -9,8 +9,9 @@ import code.graphics.SpriteData;
  */
 public class ParticleComponent extends Component {
     private SpriteData particle;
-    private int scale;
-    public ParticleComponent(SpriteData particle, int scale) {
+    private float scale;
+
+    public ParticleComponent(SpriteData particle, float scale) {
         super(ComponentType.PARTICLE);
         this.particle = particle;
         this.scale = scale;
@@ -24,11 +25,11 @@ public class ParticleComponent extends Component {
         this.particle = particle;
     }
 
-    public int getScale() {
+    public float getScale() {
         return scale;
     }
 
-    public void setScale(int scale) {
+    public void setScale(float scale) {
         this.scale = scale;
     }
 }
